Added docker template that has separate test job
The original docker pipeline template was moved to docker/base.  The
difference between base and ext-tests is that base expects testing to be
done in the Dockerfile build, and ext-tests has a separate test phase
that runs in the edge Docker image, mounts a test repo in that image
(can be the same repo as the source, or a different one) and runs a
specified command.